refactor. 기존의 뉴스 생성 스케줄러 리팩토링
